[z1258] Small code style fix backporting.
(based on master commit 836ff4f)

(based on commit a70c577)

+++ src/game/Group.cpp
@@ -843,8 +843,8 @@ void Group::SetTargetIcon(uint8 id, ObjectGuid targetGuid)
     m_targetIcons[id] = targetGuid;
 
     WorldPacket data(MSG_RAID_TARGET_UPDATE, (2+8));
-    data << (uint8)0;
-    data << id;
+    data << uint8(0);
+    data << uint8(id);
     data << targetGuid;
     BroadcastPacket(&data, true);
 }
